+++ ReleaseNotes.md
@@ -5,6 +5,7 @@
 * Fixed problem with incompatible data types for some data providers such as Oracle
 * Fixed problem with using required flag on fields in composite and collection properties
 * Fixed problem with hidden publish and save buttons on iOS
+* Fixed problem with freezing page tree when moving pages in admin
 
 ### New in 1.2.3
 * Fixed script reference problem for property editors in Chrome